requirements.prod.txt
beautifulsoup4==4.12.2
bs4==0.0.1
ailab-db @ git+https://github.com/ai-cfia/ailab-db@main
lxml==4.9.3
numpy==1.25.1
openai==0.27.8
psycopg==3.1.9
psycopg-binary==3.1.9
Scrapy==2.9.0
tiktoken==0.4.0